Refactor error handling in BinaryOptionsToolsError; update websocket connection error type
- Changed WebsocketConnectionError to use Box<TungsteniteError> for better error handling.
- Added From<TungsteniteError> implementation for BinaryOptionsToolsError.
- Cleaned up imports and improved error messages for clarity.

Remove unused imports in send.rs

- Removed the unused `debug` import from tracing.

Refactor RecieverStream and FilteredRecieverStream in stream.rs

- Cleaned up code formatting and improved readability.
- Ensured consistent use of async/await patterns.
- Updated error handling in receive methods.

Update traits in traits.rs

- Cleaned up imports and ensured consistent formatting.
- Maintained trait definitions for better clarity and usability.

Refactor Data and Callback structures in types.rs

- Improved code organization and readability.
- Ensured consistent use of async patterns and error handling.
- Updated the default_validator function for clarity.

Update reimports in reimports.rs

- Organized imports for better readability.

Bump binary-options-tools-macros version to 0.2.0

- Updated version in Cargo.lock to reflect the latest changes.

+++ BinaryOptionsToolsUni/src/platforms/pocketoption/types.rs
@@ -321,7 +321,7 @@ impl From<OriginalCandle> for Candle {
     fn from(candle: OriginalCandle) -> Self {
         Self {
             symbol: candle.symbol,
-            timestamp: candle.timestamp as i64,
+            timestamp: candle.timestamp,
             open: candle.open.to_f64().unwrap_or_default(),
             high: candle.high.to_f64().unwrap_or_default(),
             low: candle.low.to_f64().unwrap_or_default(),
